Harden frontend install locking and cache keys
